Organize `nnbench.types.types` into different files
such that the file is less long. The files aim to group related
types into files with fitting names.

Also refactored imports around the projects and in tests accordingly.
