+++ src/nnbench/types/__init__.py
@@ -1,1 +1,2 @@
-from .types import Benchmark, BenchmarkRecord, Memo, Parameters, State
+from .benchmark import Benchmark, BenchmarkRecord, Parameters, State
+from .memo import Memo
