--- src/nnbench/types/types.py
@@ -1,300 +1,0 @@
-"""Useful type interfaces to override/subclass in benchmarking workflows."""
-
-from __future__ import annotations
-
-import copy
-import functools
-import inspect
-import logging
-import threading
-from dataclasses import dataclass, field
-from types import MappingProxyType
-from typing import Any, Callable, Generic, Literal, Mapping, Protocol, TypeVar
-
-from nnbench.context import Context
-
-T = TypeVar("T")
-Variable = tuple[str, type, Any]
-
-_memo_cache: dict[int, Any] = {}
-_cache_lock = threading.Lock()
-
-logger = logging.getLogger(__name__)
-
-
-def memo_cache_size() -> int:
-    """
-    Get the current size of the memo cache.
-
-    Returns
-    -------
-    int
-        The number of items currently stored in the memo cache.
-    """
-    return len(_memo_cache)
-
-
-def clear_memo_cache() -> None:
-    """
-    Clear all items from memo cache in a thread_safe manner.
-    """
-    with _cache_lock:
-        _memo_cache.clear()
-
-
-def evict_memo(_id: int) -> Any:
-    """
-    Pop cached item with key `_id` from the memo cache.
-
-    Parameters
-    ----------
-    _id : int
-        The unique identifier (usually the id assigned by the Python interpreter) of the item to be evicted.
-
-    Returns
-    -------
-    Any
-        The value that was associated with the removed cache entry. If no item is found with the given `_id`, a KeyError is raised.
-    """
-    with _cache_lock:
-        return _memo_cache.pop(_id)
-
-
-def cached_memo(fn: Callable) -> Callable:
-    """
-    Decorator that caches the result of a method call based on the instance ID.
-
-    Parameters
-    ----------
-    fn: Callable
-        The method to memoize.
-
-    Returns
-    -------
-    Callable
-        A wrapped version of the method that caches its result.
-    """
-
-    @functools.wraps(fn)
-    def wrapper(self, *args, **kwargs):
-        _tid = id(self)
-        with _cache_lock:
-            if _tid in _memo_cache:
-                logger.debug(f"Returning memoized value from cache with ID {_tid}")
-                return _memo_cache[_tid]
-        logger.debug(f"Computing value on memo with ID {_tid} (cache miss)")
-        value = fn(self, *args, **kwargs)
-        with _cache_lock:
-            _memo_cache[_tid] = value
-        return value
-
-    return wrapper
-
-
-def NoOp(state: State, params: Mapping[str, Any] = MappingProxyType({})) -> None:
-    pass
-
-
-class CallbackProtocol(Protocol):
-    def __call__(self, state: State, params: Mapping[str, Any]) -> None: ...
-
-
-@dataclass(frozen=True)
-class BenchmarkRecord:
-    context: Context
-    benchmarks: list[dict[str, Any]]
-
-    def compact(
-        self,
-        mode: Literal["flatten", "inline", "omit"] = "inline",
-        sep: str = ".",
-    ) -> list[dict[str, Any]]:
-        """
-        Prepare the benchmark results, optionally inlining the context either as a
-        nested dictionary or in flattened form.
-
-        Parameters
-        ----------
-        mode: Literal["flatten", "inline", "omit"]
-            How to handle the context. ``"omit"`` leaves out the context entirely, ``"inline"``
-            inserts it into the benchmark dictionary as a single entry named ``"context"``, and
-            ``"flatten"`` inserts the flattened context values into the dictionary.
-        sep: str
-            The separator to use when flattening the context, i.e. when ``mode = "flatten"``.
-
-        Returns
-        -------
-        list[dict[str, Any]]
-            The updated list of benchmark records.
-        """
-        if mode == "omit":
-            return self.benchmarks
-
-        result = []
-
-        for b in self.benchmarks:
-            bc = copy.deepcopy(b)
-            if mode == "inline":
-                bc["context"] = self.context.data
-            elif mode == "flatten":
-                flat = self.context.flatten(sep=sep)
-                bc.update(flat)
-                bc["_contextkeys"] = list(self.context.keys())
-            result.append(bc)
-        return result
-
-    @classmethod
-    def expand(cls, bms: list[dict[str, Any]]) -> BenchmarkRecord:
-        """
-        Expand a list of deserialized JSON-like objects into a benchmark record.
-        This is equivalent to extracting the context given by the method it was
-        serialized with, and then returning the rest of the data as is.
-
-        Parameters
-        ----------
-        bms: list[dict[str, Any]]
-            The list of benchmark dicts to expand into a record.
-
-        Returns
-        -------
-        BenchmarkRecord
-            The resulting record with the context extracted.
-
-        """
-        dctx: dict[str, Any] = {}
-        for b in bms:
-            if "context" in b:
-                dctx = b.pop("context")
-            elif "_contextkeys" in b:
-                ctxkeys = b.pop("_contextkeys")
-                for k in ctxkeys:
-                    # This should never throw, save for data corruption.
-                    dctx[k] = b.pop(k)
-        return cls(context=Context.make(dctx), benchmarks=bms)
-
-    # TODO: Add an expandmany() API for returning a sequence of records for heterogeneous
-    #  context data.
-
-
-@dataclass(frozen=True)
-class State:
-    name: str
-    family: str
-    family_size: int
-    family_index: int
-
-
-class Memo(Generic[T]):
-    """Abstract base class for memoized values in benchmark runs."""
-
-    # TODO: Make this better than the decorator application
-    #  -> _Cached metaclass like in fsspec's AbstractFileSystem (maybe vendor with license)
-
-    @cached_memo
-    def __call__(self) -> T:
-        """Placeholder to override when subclassing. The call should return the to be cached object."""
-        raise NotImplementedError
-
-    def __del__(self) -> None:
-        """Delete the cached object and clear it from the cache."""
-        with _cache_lock:
-            sid = id(self)
-            if sid in _memo_cache:
-                logger.debug(f"Deleting cached value for memo with ID {sid}")
-                del _memo_cache[sid]
-
-
-@dataclass(init=False, frozen=True)
-class Parameters:
-    """
-    A dataclass designed to hold benchmark parameters. This class is not functional
-    on its own, and needs to be subclassed according to your benchmarking workloads.
-
-    The main advantage over passing parameters as a dictionary is, of course,
-    static analysis and type safety for your benchmarking code.
-    """
-
-    pass
-
-
-@dataclass(frozen=True)
-class Benchmark:
-    """
-    Data model representing a benchmark. Subclass this to define your own custom benchmark.
-
-    Parameters
-    ----------
-    fn: Callable[..., Any]
-        The function defining the benchmark.
-    name: str | None
-        A name to display for the given benchmark. If not given, will be constructed from the
-        function name and given parameters.
-    params: dict[str, Any]
-        A partial parametrization to apply to the benchmark function. Internal only,
-        you should not need to set this yourself.
-    setUp: Callable[..., None]
-        A setup hook run before the benchmark. Must take all members of `params` as inputs.
-    tearDown: Callable[..., None]
-        A teardown hook run after the benchmark. Must take all members of `params` as inputs.
-    tags: tuple[str, ...]
-        Additional tags to attach for bookkeeping and selective filtering during runs.
-    """
-
-    fn: Callable[..., Any]
-    name: str = field(default="")
-    params: dict[str, Any] = field(default_factory=dict)
-    setUp: Callable[[State, Mapping[str, Any]], None] = field(repr=False, default=NoOp)
-    tearDown: Callable[[State, Mapping[str, Any]], None] = field(repr=False, default=NoOp)
-    tags: tuple[str, ...] = field(repr=False, default=())
-    interface: Interface = field(init=False, repr=False)
-
-    def __post_init__(self):
-        if not self.name:
-            super().__setattr__("name", self.fn.__name__)
-        super().__setattr__("interface", Interface.from_callable(self.fn, self.params))
-
-
-@dataclass(frozen=True)
-class Interface:
-    """
-    Data model representing a function's interface. An instance of this class
-    is created using the `from_callable` class method.
-
-    Parameters:
-    ----------
-    names : tuple[str, ...]
-        Names of the function parameters.
-    types : tuple[type, ...]
-        Types of the function parameters.
-    defaults : tuple
-        A tuple of the function parameters' default values.
-    variables : tuple[Variable, ...]
-        A tuple of tuples, where each inner tuple contains the parameter name and type.
-    returntype: type
-        The function's return type annotation, or NoneType if left untyped.
-    """
-
-    names: tuple[str, ...]
-    types: tuple[type, ...]
-    defaults: tuple
-    variables: tuple[Variable, ...]
-    returntype: type
-
-    @classmethod
-    def from_callable(cls, fn: Callable, defaults: dict[str, Any]) -> Interface:
-        """
-        Creates an interface instance from the given callable.
-        """
-        # Set `follow_wrapped=False` to get the partially filled interfaces.
-        # Otherwise we get missing value errors for parameters supplied in benchmark decorators.
-        sig = inspect.signature(fn, follow_wrapped=False)
-        ret = sig.return_annotation
-        _defaults = {k: defaults.get(k, v.default) for k, v in sig.parameters.items()}
-        # defaults are the signature parameters, then the partial parametrization.
-        return cls(
-            tuple(sig.parameters.keys()),
-            tuple(p.annotation for p in sig.parameters.values()),
-            tuple(_defaults.values()),
-            tuple((k, v.annotation, _defaults[k]) for k, v in sig.parameters.items()),
-            type(ret) if ret is None else ret,
-        )
